+++ infrastructure/entities/user.go
@@ -1,9 +1,7 @@
 package entities
 
 import (
-	"golang.org/x/crypto/bcrypt"
 	"gorm.io/gorm"
-	"log"
 )
 
 type User struct {
@@ -11,13 +9,4 @@ type User struct {
 	Username string `gorm:"not null"`
 	Email    string `gorm:"not null"`
 	Password string `gorm:"not null"`
-}
-
-func Salt_pwd(pwd string) string {
-	hash, err := bcrypt.GenerateFromPassword([]byte(pwd), bcrypt.MinCost)
-	if err != nil {
-		log.Fatal("The given password could not be hashed. ", err)
-	}
-
-	return string(hash)
 }
